Moved settings file into new settings package __init__
I really need to resolve the project structure....
